修复全部编译警告: 项目代码 + 第三方库 (除 MSB8012)
本次清除 4 个配置(Win32/x64 x Debug/Release)日志中的全部 warning, 仅保留 MSB8012 工程配置警告。

第三方库(skia/mhook) 39处 - vcxproj DisableSpecificWarnings:

- skia/skia64: 追加 4267;4312;4334 (SkStream/SkDescriptor/GrDrawTarget/GrGpu/SkEdge 等)

- mhook/mhook64: 追加 4477;4474;5287;4313 (disasm_x86.c printf 格式/枚举)

- render-skia/render-skia64: 追加 4244 (SkFloatingPoint.h copysign double->float)

- render-gdi/render-gdi64: 追加 4244 (一致性)

第三方单文件库 stb_image.h 57处 C4312 - #pragma 局部屏蔽 include (不改第三方源码)

项目代码 147处 - 逐点精确修复:

C4267 (size_t->int) ~46处: render-skia/drawtext-skia/log4z/render-gdi/SChromeTabCtrl/pugixml/FileHelper/QR_Encode/spropertyitembase/SHeaderCtrlEx/SPathView/imgdecoder-stb 等, 加 (int) 显式转换

C4244 (LPARAM/WPARAM/ULONG_PTR/LONG_PTR->目标) ~37处: SMCListViewEx(15)/SwkeWebkit/MainDlg/SimpleWnd/shostwnd/SWndFinder/SIpcObject/slistctrlex/SMCListViewEx.h(C4244 WPARAM->char, 修 SwndMsgCracker.h MSG_WM_TIMER2 宏加 cast) 等

C4312/C4311/C4302 指针截断 4处 (genericserver.cpp) - 真实bug修复: (DWORD)ptr -> (ULONG_PTR)ptr (x64下指针被截断为32位, 导致HTTP响应数据写入错误地址)

C4477 格式串 1处 (httpserver.cpp): Content-Length 用 %Iu + (size_t) (MSVC原生size_t格式符)

C4244/C4267 adapter.h/CAdapter.h/SMCListViewEx.cpp GetItemIndex/GetUserData 12处: 加 (int)/(long) cast

编码处理: utf8/utf8-bom/gbk/utf16-le 五种编码均按原格式写回 (统一编码感知脚本)

修复过程中纠正的自身错误: SMCListViewEx.cpp 批量replace在链式调用 GetHead().pItem 上 (long) 错位导致 C2059 语法错误, 改用整行精确替换; pugixml (int) cast 位置错误(加在外层返回值而非参数), 纠正到传参处。

+++ SOUI/include/helper/SwndFinder.h
@@ -43,7 +43,7 @@ namespace SOUI
 			key.sep.findByName=fi.findByName;
 
 			lRet += key.key;
-			return lRet;
+			return (ULONG)lRet;
 		}
 
 		static bool CompareElements(INARGTYPE element1, INARGTYPE element2)
