+++ components/render-skia/drawtext-skia.cpp
@@ -28,7 +28,7 @@ static size_t breakTextEx(const SkPaint *pPaint, const wchar_t* textD, size_t le
 
 SkRect DrawText_Skia(SkCanvas* canvas,const wchar_t *text,int len,SkRect box,const SkPaint& paint,UINT uFormat)
 {
-	if(len<0)	len = wcslen(text);
+	if(len<0)	len = (int)wcslen(text);
     SkTextLayoutEx layout;
     layout.init(text,len,box,paint,uFormat);
     
@@ -40,13 +40,13 @@ void SkTextLayoutEx::init( const wchar_t text[], size_t length,SkRect rc, const 
 {
     if(uFormat & DT_NOPREFIX)
     {
-        m_text.setCount(length);
+        m_text.setCount((int)length);
         memcpy(m_text.begin(),text,length*sizeof(wchar_t));
     }else
     {
         m_prefix.reset();
         SkTDArray<wchar_t> tmp;
-        tmp.setCount(length);
+        tmp.setCount((int)length);
         memcpy(tmp.begin(),text,length*sizeof(wchar_t));
         for(int i=0;i<tmp.count();i++)
         {
@@ -92,7 +92,7 @@ void SkTextLayoutEx::buildLines()
             size_t line_len = breakTextEx(m_paint,text, stop - text, maxWid,0);
 			if (0 == line_len) break;
 			text += line_len;
-            lineHead += line_len;
+            lineHead += (int)line_len;
         };
     }
 }
